Seed and troubleshoot

+++ server/routes/team_by_id.py
@@ -9,33 +9,34 @@ teams_schema = TeamSchema(many=True, session=db.session)
 
 
 class TeamById(Resource):
-    def get(self, team_id):
+    def get(self, id):
         try:
-            return team_schema.dump(Team.query.get(team_id))
+            if team := db.session.get(Team, id):
+                return team_schema.dump(team)
+            return {"message": "Team not found"}, 404
         except Exception as e:
             return {"message": str(e)}, 500
 
-    def patch(self, team_id):
-        try:
-            data = request.json
-            team_schema.validate(data)
-            team = Team.query.get(team_id)
-            team.name = data["name"]
-            team.city = data["city"]
-            team.state = data["state"]
-            team.league_id = data["league_id"]
-            db.session.commit()
-            return team_schema.dump(team), 200
-        except Exception as e:
-            db.session.rollback()
-            return {"message": str(e)}, 400
+    def patch(self, id):
+        if team := db.session.get(Team, id):
+            try:
+                data = request.json
+                team_schema.validate(data)
+                team = team_schema.load(data, instance=team, partial=True)
+                db.session.commit()
+                return team_schema.dump(team)
+            except Exception as e:
+                db.session.rollback()
+                return {"message": str(e)}, 400
+        return {"message": "Team not found"}, 404
 
-    def delete(self, team_id):
-        try:
-            team = Team.query.get(team_id)
-            db.session.delete(team)
-            db.session.commit()
-            return {"message": "Team deleted"}, 204
-        except Exception as e:
-            db.session.rollback()
-            return {"message": str(e)}, 400
+    def delete(self, id):
+        if team := db.session.get(Team, id):
+            try:
+                db.session.delete(team)
+                db.session.commit()
+                return {"message": f"Team {id} deleted"}
+            except Exception as e:
+                db.session.rollback()
+                return {"message": str(e)}, 400
+        return {"message": "Team not found"}, 404
